pyproject.toml
[tool.poetry]
name = "data-point"
version = "0.1.0"
description = ""
authors = ["Noat"]
readme = "README.md"

packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "0.103.2"
click = "8.1.7"
uvicorn = "0.23.2"
sqlalchemy = {extras = ["asyncio"], version = "2.0.21"}
pydantic-settings = "2.0.3"
alembic = "1.12.0"
asyncpg = "0.28.0"
flake8 = "6.1.0"
jinja2 = "3.1.2"
mypy = "^1.10.0"
google-api-python-client = "^2.127.0"

[tool.poetry.scripts]
app = "src.cli.app:app_start"

[tool.mypy]
python_version = "3.10"
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin"
]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
show_column_numbers = true
show_error_codes = true
no_implicit_optional = true
pretty = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true
warn_unreachable = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
